Move models into a shared directory.
