+++ lsif/src/server/server.ts
@@ -21,7 +21,7 @@ import { defineRedisCommands } from './redis/redis'
 import { errorHandler } from './middleware/errors'
 import { logger as loggingMiddleware } from 'express-winston'
 import { Logger } from 'winston'
-import { LsifDump } from '../xrepo.models'
+import { LsifDump } from '../shared/models/xrepo'
 import { metricsMiddleware } from './middleware/metrics'
 import { pipeline as _pipeline } from 'stream'
 import { queueSizeGauge } from '../server.metrics'
