+++ lsif/src/shared/database/postgres.ts
@@ -1,5 +1,5 @@
 import { Connection, createConnection as _createConnection } from 'typeorm'
-import * as xrepo from '../../xrepo.models'
+import * as xrepo from '../models/xrepo'
 import { PostgresConnectionCredentialsOptions } from 'typeorm/driver/postgres/PostgresConnectionCredentialsOptions'
 import { readEnvInt } from '../../util'
 import { Logger } from 'winston'
