+++ lsif/src/shared/models/types.ts
