+++ lsif/src/shared/xrepo/xrepo.ts
@@ -8,7 +8,7 @@ import * as crc32 from 'crc-32'
 import { instrument } from '../metrics'
 import { Connection, EntityManager, Brackets } from 'typeorm'
 import { createFilter, testFilter } from '../../encoding'
-import { PackageModel, ReferenceModel, Commit, LsifDump, DumpId } from '../../xrepo.models'
+import { PackageModel, ReferenceModel, Commit, LsifDump, DumpId } from '../models/xrepo'
 import { TableInserter } from '../database/inserter'
 import { addrFor, getCommitsNear, gitserverExecLines } from './commits'
 import { TracingContext, logAndTraceCall } from '../tracing'
