+++ lsif/src/worker/importer/correlator.ts
@@ -2,7 +2,7 @@ import RelateUrl from 'relateurl'
 import { mustGet, mustGetFromEither } from '../../util'
 import { DefaultMap } from '../../default-map'
 import { Hover, MarkupContent } from 'vscode-languageserver-types'
-import { MonikerData, PackageInformationData, RangeData } from '../../database.models'
+import { MonikerData, PackageInformationData, RangeData } from '../../shared/models/dump'
 import {
     MonikerId,
     HoverResultId,
@@ -10,7 +10,7 @@ import {
     DefinitionResultId,
     DocumentId,
     PackageInformationId,
-} from '../../database.types'
+} from '../../shared/models/types'
 import {
     Id,
     VertexLabels,
